Refactor to use package oriented design
