+++ .covignore
@@ -1,0 +1,2 @@
+src/cmd/controller/add_image.go
+src/databases/
