--- cmd/api/main.go
@@ -1,7 +1,0 @@
-package main
-
-import "github.com/sanrinconr/storj-images/cmd/server"
-
-func main() {
-	server.Start()
-}
