--- cmd/controller/ping_test.go
@@ -1,20 +1,0 @@
-package controller_test
-
-import (
-	"net/http/httptest"
-	"testing"
-
-	"github.com/gin-gonic/gin"
-	"github.com/sanrinconr/storj-images/cmd/controller"
-	"github.com/stretchr/testify/assert"
-)
-
-func TestPing(t *testing.T) {
-	gin.SetMode(gin.TestMode)
-	c, _ := gin.CreateTestContext(httptest.NewRecorder())
-	ctl := controller.Ping{}
-
-	got := ctl.Ping(c)
-
-	assert.Nil(t, got)
-}
