+++ docker-compose.yml
@@ -4,7 +4,7 @@ services:
 
   backend:
     container_name: storj-images-backend-container
-    image: ${REGISTRY}/storj-images-backend
+    image: registry.santiagorincon.tk/storj-images-backend
     build: 
       context: ./
       dockerfile: Dockerfile
