+++ src/cmd/dependencies/packages.go
@@ -1,0 +1,40 @@
+package dependencies
+
+import (
+	"time"
+
+	"github.com/sanrinconr/storj-images/src/cmd/config"
+	"github.com/sanrinconr/storj-images/src/upload"
+)
+
+type packages struct {
+	config config.Config
+	timer  func() time.Time
+}
+
+func newPackages(c config.Config, opts ...func() time.Time) packages {
+	timer := defaultTimer()
+	if len(opts) > 0 && opts[0] != nil {
+		timer = opts[0]
+	}
+
+	return packages{c, timer}
+}
+
+func (p packages) uploadAddImage() upload.AddImage {
+	r, err := upload.NewRepository(Mongo(p.config), Storj(p.config), defaultTimer())
+	if err != nil {
+		panic(err)
+	}
+
+	u, err := upload.NewAddImage(r, p.config.AllowedFormats, p.timer)
+	if err != nil {
+		panic(err)
+	}
+
+	return u
+}
+
+func defaultTimer() func() time.Time {
+	return time.Now
+}
