+++ src/cmd/main.go
@@ -1,0 +1,8 @@
+// Package main are the entrypoint for setup the api and run the daemon.
+package main
+
+import "github.com/sanrinconr/storj-images/src/cmd/api"
+
+func main() {
+	api.Start()
+}
